# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/digest_digests:
    get:
      tags:
        - DigestDigests
      summary: List DigestDigests
      description: Get a paginated list of DigestDigests with optional filtering and sorting
      operationId: listDigestDigests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestDigestList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - DigestDigests
      summary: Create DigestDigest
      description: Create a new DigestDigest
      operationId: createDigestDigest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDigestDigestRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestDigest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/digest_digest_kpis:
    get:
      tags:
        - DigestDigestKpis
      summary: List DigestDigestKpis
      description: Get a paginated list of DigestDigestKpis with optional filtering and sorting
      operationId: listDigestDigestKpis
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestDigestKpiList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - DigestDigestKpis
      summary: Create DigestDigestKpi
      description: Create a new DigestDigestKpi
      operationId: createDigestDigestKpi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDigestDigestKpiRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestDigestKpi'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    DigestDigest:
      type: object
      required:
        - id
        - name
        - periodicity
        - state
        - company_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 255
        periodicity:
          type: string
        next_run_date:
          nullable: true
          type: string
          format: date
        state:
          type: string
        company_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateDigestDigestRequest:
      type: object
      required:
        - name
        - periodicity
        - state
        - company_id
        - metadata
      properties:
        name:
          type: string
          maxLength: 255
        periodicity:
          type: string
        next_run_date:
          nullable: true
          type: string
          format: date
        state:
          type: string
        company_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    DigestDigestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DigestDigest'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    DigestDigestKpi:
      type: object
      required:
        - id
        - digest_id
        - kpi_key
        - metadata
      properties:
        id:
          type: string
          format: uuid
        digest_id:
          type: string
          format: uuid
        kpi_key:
          type: string
          maxLength: 120
        metadata:
          type: object
...
    CreateDigestDigestKpiRequest:
      type: object
      required:
        - digest_id
        - kpi_key
        - metadata
      properties:
        digest_id:
          type: string
          format: uuid
        kpi_key:
          type: string
          maxLength: 120
        metadata:
          type: object
...
    DigestDigestKpiList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DigestDigestKpi'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []